Complete rework
The application has been completely reworked, it now uses the new WPForms instead of the old WinForms, it is now much faster to open and use. Additionally, it is much simpler written, and is a lot cleaner and straight forward to use.

Comments have been added extensively throughout the project in order to help any contributors (or even future me!) to understand what is happening at any given point.